backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
playwright==1.41.2
playwright-stealth==1.0.6
requests==2.31.0
httpx==0.26.0
python-dotenv==1.0.1
# LLM
openai==1.12.0
langchain==0.1.6
langchain-openai==0.0.5
# 单元测试
pytest==8.0.0
pytest-asyncio==0.23.5
# 数据库
sqlalchemy==2.0.27
# 调度
apscheduler==3.10.4
# 文件解析
python-docx==1.1.0
